pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "flask>=3.1.1",
    "google-generativeai>=0.8.5",
    "psutil>=7.0.0",
    "replit>=4.1.2",
]
